refactor(verticaltabs): css import paths
