+++ packages/verticaltabs/src/index.ts
@@ -1,1 +1,3 @@
-export { default } from './react/index'
+import './css/index.css'
+import VerticalTabs from './react/index'
+export default VerticalTabs
