+++ packages/verticaltabs/src/react/divider.tsx
@@ -2,8 +2,6 @@ import { useTheme } from '@pluralsight/ps-design-system-theme'
 import { classNames } from '@pluralsight/ps-design-system-util'
 import React from 'react'
 
-import '../css/index.css'
-
 const Divider: React.FC<React.HTMLAttributes<HTMLHRElement>> = ({
   className,
   ...props
